requirements.txt
opencv-python
scikit-image
streamlit
faiss-cpu
torch
torchvision